fixing ddl bugs

+++ db/tables/latest_balances.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS latest_balances(
-    accountNumber INTEGER REFERENCES account(accountNumber),
+    accountNumber NUMERIC(16, 0) REFERENCES account(accountNumber),
     amount NUMERIC(18, 0)
 );
